changed unsigned int i to size_t i

+++ search_algorithms/0-linear.c
@@ -11,7 +11,7 @@
  */
 int linear_search(int *array, size_t size, int value)
 {
-	unsigned int i = 0;
+	size_t i = 0;
 
 	if (array[i] == '\0')
 	{
@@ -19,7 +19,7 @@ int linear_search(int *array, size_t size, int value)
 	}
 	for (i = 0; i < size; i++)
 	{
-		printf("value checked array[%u] = [%d]\n", i, array[i]);
+		printf("value checked array[%lu] = [%d]\n", i, array[i]);
 		if (array[i] == value)
 		{
 			return (i);
